feat: 회원가입 리팩토링 (#ATR-147)
