+++ apps/service/entities/auth/constant/path.ts
@@ -1,3 +1,3 @@
-export const PUBLIC_PATH = ['/sign-in']
+export const PUBLIC_PATH = ['/sign-in', '/sign-up']
 
 export const PRIVATE_PATH = ['/inbox', '/mypage']
