+++ apps/service/features/sign-up/api/index.ts
@@ -1,0 +1,1 @@
+export { default as postDuplicateName } from './postDuplicateInput'
