+++ apps/service/features/sign-up/constant/birthRegex.ts
@@ -1,0 +1,3 @@
+const BIRTH_REGEX = /^(19|20)\d{2}(0[1-9]|1[0-2])(0[1-9]|[12]\d|3[01])$/
+
+export default BIRTH_REGEX
