+++ apps/service/features/sign-up/constant/index.ts
@@ -1,0 +1,1 @@
+export { default as BIRTH_REGEX } from './birthRegex'
