+++ apps/service/features/sign-up/lib/hook/index.ts
@@ -1,0 +1,1 @@
+export { default as useDisabledBtn } from './useDisabledBtn'
