+++ apps/service/features/sign-up/lib/index.ts
@@ -1,0 +1,1 @@
+export * from './hook'
